# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bp-carnet",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/account/email": {
      "post": {
        "summary": "新しいメールアドレスに、変更を確かめるリンクを送る。",
        "description": "パスワードを持つアカウントは、今のパスワードを求める (パスワードの変更と同じく、セッションだけを\n奪った相手に変えられないようにするため)。応答は新しいアドレスが使われているかによらず同じにし、\n違いはメールの中身だけにする (サインアップと同じく、アドレスから利用の有無を探らせないため)。",
        "operationId": "request_email_change",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestEmailChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "受け付けた (新しいアドレスが使われていても同じ応答)"
          },
          "400": {
            "description": "リクエストボディが不正 (JSON として解釈できない等)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未ログイン",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Content-Type が application/json でない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "メールアドレスとして解釈できない・今のアドレスと同じ・今のパスワードが一致しない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "要求か、今のパスワードの照合の試行回数が上限を超えた",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/email-change/check": {
      "post": {
        "summary": "変更のリンクがまだ使えれば、変更先のアドレスを返す (リンクを開いたときに見せて確かめてもらう)。",
        "operationId": "check_email_change_link",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "リンクは使える",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailChangeTargetResponse"
                }
              }
            }
          },
          "400": {
            "description": "リンクが無効・期限切れ・使用済み、またはリクエストボディが不正",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Content-Type が application/json でない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "必須フィールドが無い",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/email-change/complete": {
      "post": {
        "summary": "リンクの先で変更を確定し、古いアドレスに知らせる。ログインしていなくても確定できる\n(リンクが新しいアドレスを受け取れることの確かめになるため。サインアップの確認と同じ)。",
        "operationId": "complete_email_change",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "変更した"
          },
          "400": {
            "description": "リンクが無効・期限切れ・使用済み、またはリクエストボディが不正",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "新しいアドレスが、申し込みの後に別のアカウントで使われた",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Content-Type が application/json でない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "必須フィールドが無い",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckLinkRequest": {
        "type": "object",
        "description": "リンクの先を開いたときに、リンクがまだ使えるかを尋ねる要求。",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string",
            "description": "メールのリンクに載っていたトークン。"
          }
        }
      },
...
      "EmailChangeTargetResponse": {
        "type": "object",
        "required": [
          "newEmail"
        ],
        "properties": {
          "newEmail": {
            "type": "string",
            "description": "変更先のメールアドレス。"
          }
        }
      },
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "rows": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/RowError"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "レスポンスボディの共通 envelope。`code` が機械可読な契約 (frontend はこちらで表示文言を引く)、\n`message` はローカライズしないデバッグ用の説明文。`rows` は `BatchValidation` のときだけ埋まる。",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "RequestEmailChangeRequest": {
        "type": "object",
        "required": [
          "newEmail"
        ],
        "properties": {
          "currentPassword": {
            "type": [
              "string",
              "null"
            ],
            "description": "今のパスワード。パスワードを持つアカウントだけ必須 (持たないアカウントは送らない)。"
          },
          "newEmail": {
            "type": "string",
            "example": "new@example.com"
          }
        }
      },
...
      "RowError": {
        "type": "object",
        "description": "バッチ処理で 1 行の検証が失敗したことを表す。`index` はリクエスト配列内の 0 始まりの位置。",
        "required": [
          "index",
          "message"
        ],
        "properties": {
          "index": {
            "type": "integer",
            "minimum": 0
          },
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  }
}